refactor: PartyList View 파일 분리
